refactor: move modal element to not be child to mapSidebar__header to avoid modal not showing due to mapSidebar__header position sticky
